Changed getEstimatedTaskTime out parameters type from Float to Double.
git-svn-id: http://svn.ofbiz.org/svn/ofbiz/trunk@6256 deaff617-5be1-0310-a1b7-ec6878f92ba9

+++ applications/manufacturing/src/org/ofbiz/manufacturing/routing/RoutingServices.java
@@ -83,10 +83,10 @@ public class RoutingServices {
         long estimatedTaskTime = ProductionRun.getEstimatedTaskTime(task, quantity, productId, routingId, dispatcher);
         result.put("estimatedTaskTime", new Long(estimatedTaskTime));
         if (task != null && task.get("estimatedSetupMillis") != null) {
-            result.put("setupTime", task.getFloat("estimatedSetupMillis"));
+            result.put("setupTime", task.getDouble("estimatedSetupMillis"));
         }
         if (task != null && task.get("estimatedMilliSeconds") != null) {
-            result.put("taskUnitTime", task.getFloat("estimatedMilliSeconds"));
+            result.put("taskUnitTime", task.getDouble("estimatedMilliSeconds"));
         }
         return result;
     }
